+++ Packages/com.unity.inputsystem/InputSystem/Editor/Settings/InputSettingsProvider.cs
@@ -145,7 +145,6 @@ namespace UnityEngine.InputSystem.Editor
                 EditorGUILayout.LabelField("Editor", EditorStyles.boldLabel);
                 EditorGUILayout.Space();
                 EditorGUILayout.PropertyField(m_EditorInputBehaviorInPlayMode, m_EditorInputBehaviorInPlayModeContent);
-                EditorGUILayout.PropertyField(m_InputActionPropertyDrawerMode, m_InputActionPropertyDrawerModeContent);
 
                 EditorGUILayout.Space();
                 EditorGUILayout.LabelField("Improved Shortcut Support", EditorStyles.boldLabel);
@@ -275,7 +274,6 @@ namespace UnityEngine.InputSystem.Editor
             m_CompensateForScreenOrientation = m_SettingsObject.FindProperty("m_CompensateForScreenOrientation");
             m_BackgroundBehavior = m_SettingsObject.FindProperty("m_BackgroundBehavior");
             m_EditorInputBehaviorInPlayMode = m_SettingsObject.FindProperty("m_EditorInputBehaviorInPlayMode");
-            m_InputActionPropertyDrawerMode = m_SettingsObject.FindProperty("m_InputActionPropertyDrawerMode");
             m_DefaultDeadzoneMin = m_SettingsObject.FindProperty("m_DefaultDeadzoneMin");
             m_DefaultDeadzoneMax = m_SettingsObject.FindProperty("m_DefaultDeadzoneMax");
             m_DefaultButtonPressPoint = m_SettingsObject.FindProperty("m_DefaultButtonPressPoint");
@@ -304,10 +302,6 @@ namespace UnityEngine.InputSystem.Editor
                 + "the game.\n"
                 + "'All Device Input Always Goes To Game View' causes input to treat 'Background Behavior' exactly as in the player including devices potentially being disabled entirely while the Game View "
                 + "does not have focus. In this setting, no input from the Input System will be visible to EditorWindows.");
-            m_InputActionPropertyDrawerModeContent = new GUIContent("Input Action Property Drawer Mode", "Determines how the Inspector window displays InputActionProperty fields.\n\n" +
-                "'Compact' displays the property in a compact format, using a minimal number of lines.\n" +
-                "'Multiline Effective' displays the effective action underlying the property, using multiple lines.\n" +
-                "'Multiline Both' displays both the input action and external reference underlying the property.");
             m_DefaultDeadzoneMinContent = new GUIContent("Default Deadzone Min", "Default 'min' value for Stick Deadzone and Axis Deadzone processors.");
             m_DefaultDeadzoneMaxContent = new GUIContent("Default Deadzone Max", "Default 'max' value for Stick Deadzone and Axis Deadzone processors.");
             m_DefaultButtonPressPointContent = new GUIContent("Default Button Press Point", "The default press point used for Button controls as well as for various interactions. For button controls which have analog physical inputs, this configures how far they need to   be held down to be considered 'pressed'.");
@@ -418,7 +412,6 @@ namespace UnityEngine.InputSystem.Editor
         [NonSerialized] private SerializedProperty m_CompensateForScreenOrientation;
         [NonSerialized] private SerializedProperty m_BackgroundBehavior;
         [NonSerialized] private SerializedProperty m_EditorInputBehaviorInPlayMode;
-        [NonSerialized] private SerializedProperty m_InputActionPropertyDrawerMode;
         [NonSerialized] private SerializedProperty m_DefaultDeadzoneMin;
         [NonSerialized] private SerializedProperty m_DefaultDeadzoneMax;
         [NonSerialized] private SerializedProperty m_DefaultButtonPressPoint;
@@ -442,7 +435,6 @@ namespace UnityEngine.InputSystem.Editor
         private GUIContent m_CompensateForScreenOrientationContent;
         private GUIContent m_BackgroundBehaviorContent;
         private GUIContent m_EditorInputBehaviorInPlayModeContent;
-        private GUIContent m_InputActionPropertyDrawerModeContent;
         private GUIContent m_DefaultDeadzoneMinContent;
         private GUIContent m_DefaultDeadzoneMaxContent;
         private GUIContent m_DefaultButtonPressPointContent;